requirements.txt
asposestorage==1.0.2
einops==0.8.1
h5py==3.15.1
hydra-core==1.3.2
matplotlib==3.10.8
numpy==2.3.5
omegaconf==2.3.0
Pillow==12.0.0
pydantic==2.12.5
python-dotenv==1.2.1
scikit_learn==1.8.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
